+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## Releace Candidate 1 (20.01.2020)
+
+Welcome to first stable release of MOP!
+
+Here's what's new compared to the last Beta:
+
+### Added
+
+- Added "Go to MOP wiki"
+- Added a dialog window when opening an external link
+- SATSUMA: Toggle Physics Only is now displayed in mod report
+- Temporarily Disable Physics Toggling is now displayed in mod report
+- Added "Advanced" settings
+- Added support for Moonshine Still mod (thanks to Hawk031)
+
+### Changes
+
+- Changed the layout of mod report
+- Minor changes related to occlusion culling
+- Moved vehicles and shop items toggling to Advanced settings
+- Compatibility Manager rewrite
+
 ## Beta 1.6 (13.01.2020)
 
 ### Added